Working on Documenting & correcting code

+++ contracts/interfaces/IMainMinter.sol
@@ -1,5 +1,5 @@
 pragma solidity 0.8.15;
 
 interface IMainMinter {
-    function isFoundingRaptorCheck(uint16 tokenId) external view returns (bool);
+    function isFoundingRaptorCheck(uint256 tokenId) external view returns (bool);
 }
